SolarEnergyForecasting/WeatherDataCollector/requirements.txt
azure-functions==1.13.2
requests==2.28.2
azure-storage-blob==12.14.1